requirements.txt
faiss-cpu==1.10.0
img2table==1.4.1
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
psycopg2==2.9.10
pytesseract==0.3.13
python-dotenv==1.0.1
spacy==3.8.4
SQLAlchemy==2.0.38
streamlit==1.42.1
supervision==0.25.1
ultralytics==8.3.78